requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.72
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
